// types.ts
export type LoggedInUser = {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  phoneNumber: string | null;
  profileImageUrl: string | null;
  jobTitle?: string | null;
  employmentStatus?: string | null;
  officeLocation?: string | null;
  startDate?: string | null;
  mustChangePassword?: boolean;
  role: string;
  branchId: number | null;
  branchName: string | null;
  countryId: number | null;
  countryName: string | null;
  departmentId: number | null;
  unitId: number | null;
};